+++ GUI.py
@@ -3,33 +3,12 @@ from PIL import ImageTk, Image
 import os
 import subprocess
 import numpy as np
+import cuadrantes
 
-IMAGE_WIDTH = 390
-IMAGE_HEIGHT = 390
-GRID_SIZE = 4  # 4x4 = 16 cuadrantes
 IMAGE_PATH = "narrador.jpg"
 IMG_OUT_PATH = "cuadrante.img"
 JPG_OUT_PATH = "cuadrante.jpg"
 
-def extraerCuadrante(img_array, quadrant_num):
-    row = (quadrant_num - 1) // GRID_SIZE
-    col = (quadrant_num - 1) % GRID_SIZE
-
-    quad_h = IMAGE_HEIGHT // GRID_SIZE
-    quad_w = IMAGE_WIDTH // GRID_SIZE
-
-    start_y = row * quad_h
-    end_y = start_y + quad_h
-    start_x = col * quad_w
-    end_x = start_x + quad_w
-
-    return img_array[start_y:end_y, start_x:end_x]
-
-def guardarImg(cuadrante_array, path):
-    cuadrante_array.astype(np.uint8).tofile(path)
-
-def cargarImg(path, shape):
-    return np.fromfile(path, dtype=np.uint8).reshape(shape)
 
 def GUI():
     # Ventana y componentes
@@ -69,12 +48,12 @@ def GUI():
     cuadranteLabel.place(x=900, y=200)
 
     def mostrarCuadrante(num):
-        cuadrante = extraerCuadrante(originalArray, num)
-        guardarImg(cuadrante, IMG_OUT_PATH)
+        cuadrante = cuadrantes.extraerCuadrante(originalArray, num)
+        cuadrantes.guardarImg(cuadrante, IMG_OUT_PATH)
 
 
         # Volver a cargar desde .img
-        cargado = cargarImg(IMG_OUT_PATH, cuadrante.shape)
+        cargado = cuadrantes.cargarImg(IMG_OUT_PATH, cuadrante.shape)
         img_pil = Image.fromarray(cargado, mode='L')
         img_pil.save(JPG_OUT_PATH)
